md: add jbmd dependency, move method
