+++ tools-md/src/main/java/module-info.java
@@ -7,6 +7,9 @@ module aya.md {
   requires static org.jetbrains.annotations;
   requires kala.collection.primitive;
 
+  requires aya.jb.md;
+  requires aya.jb.md.ij;
+
   exports org.aya.literate;
   exports org.aya.literate.parser;
   exports org.aya.literate.frontmatter;
